Pushing Progress
//Work Done
Player is now able to walk to spawn and host and run with Stamina.

Rebased Classes and made them Compatible with All Child Classes that im working on soon.

//Todo
Still have some stuff todo, like adding an Offshore Money Value to the Sql Server and then making an Method on the AMS Server Manager to get that Value from the Sql Server with the Cookie of the Player and then return the requesting Player the Value effectively.

And still missing an Interactable Script and Scripts for all the Items.