Fixed Bugs
Fixed Bug which made the Game Softlock after an Player died through an Defender. And fixed Messaging System allowing players to see, if the Game is Pending to start!

Missing:

Add Some Sound Effects.

Add Shop, for Offshore Cash.

Add System to add Offshore Cash and store of Sql Db.

Add new Map and configure everything.

Add an Lobby Area for Players to wait for game to start.

Add an Camera or so for Spectating Players. (If enough Time)